requirements.txt
streamlit
pandas
torch
nltk
transformers
newspaper3k